# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ViaRah API
  version: 1.0.0
  description: |
    ViaRah backend HTTP API.

    This OpenAPI document is the source of truth for ViaRah `/api/...` endpoints
    as part of “API completeness v1”.
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/orgs/{org_id}/workflows:
    get:
      operationId: workflows__workflows_get
      tags: [workflows]
      summary: List workflows (session member or API key read)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Workflows
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  workflows:
                    type: array
                    items:
                      $ref: "#/components/schemas/Workflow"
                required: [workflows]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: workflows__workflows_post
      tags: [workflows]
      summary: Create workflow (Admin/PM session or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                stages:
                  type: array
                  items:
                    type: object
                    additionalProperties: false
                    properties:
                      name:
                        type: string
                      order:
                        type: integer
                      category:
                        type: string
                        enum: [backlog, in_progress, qa, done]
                      progress_percent:
                        type: integer
                        minimum: 0
                        maximum: 100
                      is_done:
                        type: boolean
                      is_qa:
                        type: boolean
                      counts_as_wip:
                        type: boolean
                    required: [name, order, category, progress_percent]
              required: [name, stages]
      responses:
        "200":
          description: Workflow created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  workflow:
                    $ref: "#/components/schemas/Workflow"
                  stages:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkflowStage"
                required: [workflow, stages]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/orgs/{org_id}/projects:
    get:
      operationId: work_items__projects_get
      tags: [work_items]
      summary: List projects (session member incl client or API key read)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                required: [projects]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: work_items__projects_post
      tags: [work_items]
      summary: Create project (session Admin/PM/Member or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                description:
                  type: string
                progress_policy:
                  type: string
                  nullable: true
                  enum: [subtasks_rollup, workflow_stage]
              required: [name]
      responses:
        "200":
          description: Project created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
                required: [project]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/orgs/{org_id}/projects/{project_id}:
    get:
      operationId: work_items__project_get
      tags: [work_items]
      summary: Get project (session member incl client or API key read)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
                required: [project]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: work_items__project_patch
      tags: [work_items]
      summary: Update project (session Admin/PM/Member or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                workflow_id:
                  allOf:
                    - $ref: "#/components/schemas/Uuid"
                  nullable: true
                progress_policy:
                  type: string
                  enum: [subtasks_rollup, workflow_stage]
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
                required: [project]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: work_items__project_delete
      tags: [work_items]
      summary: Delete project (session Admin/PM/Member or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/orgs/{org_id}/projects/{project_id}/epics:
    get:
      operationId: work_items__project_epics_get
      tags: [work_items]
      summary: List project epics (session Admin/PM/Member or API key read)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Epics (with computed progress)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  epics:
                    type: array
                    items:
                      $ref: "#/components/schemas/Epic"
                required: [epics]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: work_items__project_epics_post
      tags: [work_items]
      summary: Create epic (session Admin/PM/Member or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                description:
                  type: string
                progress_policy:
                  type: string
                  nullable: true
                  enum: [subtasks_rollup, workflow_stage]
              required: [title]
      responses:
        "200":
          description: Epic created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  epic:
                    $ref: "#/components/schemas/Epic"
                required: [epic]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/orgs/{org_id}/epics/{epic_id}/tasks:
    post:
      operationId: work_items__epic_tasks_post
      tags: [work_items]
      summary: Create task in an epic (session Admin/PM/Member or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/EpicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                workflow_stage_id:
                  allOf:
                    - $ref: "#/components/schemas/Uuid"
                  nullable: true
                progress_policy:
                  type: string
                  nullable: true
                  enum: [subtasks_rollup, workflow_stage]
                start_date:
                  type: string
                  format: date
                  nullable: true
                end_date:
                  type: string
                  format: date
                  nullable: true
              required: [title]
      responses:
        "200":
          description: Task created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  task:
                    $ref: "#/components/schemas/Task"
                required: [task]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/orgs/{org_id}/tasks/{task_id}/subtasks:
    get:
      operationId: work_items__task_subtasks_get
      tags: [work_items]
      summary: List subtasks for a task (session member incl client or API key read)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/TaskId"
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Subtasks
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  subtasks:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subtask"
                required: [subtasks]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: work_items__task_subtasks_post
      tags: [work_items]
      summary: Create subtask (session Admin/PM/Member or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                description:
                  type: string
                status:
                  type: string
                start_date:
                  type: string
                  format: date
                  nullable: true
                end_date:
                  type: string
                  format: date
                  nullable: true
              required: [title]
      responses:
        "200":
          description: Subtask created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  subtask:
                    $ref: "#/components/schemas/Subtask"
                required: [subtask]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/orgs/{org_id}/templates:
    get:
      operationId: templates__templates_get
      tags: [templates]
      summary: List templates (session member or API key read)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  templates:
                    type: array
                    items:
                      $ref: "#/components/schemas/Template"
                required: [templates]
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: templates__templates_post
      tags: [templates]
      summary: Create template (Admin/PM session or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Template created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  template:
                    $ref: "#/components/schemas/Template"
                required: [template]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/orgs/{org_id}/report-runs:
    get:
      operationId: reports__report_runs_get
      tags: [reports]
      summary: List report runs for a project (session member or API key read)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - name: project_id
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: Report runs
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  report_runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportRun"
                required: [report_runs]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: reports__report_runs_post
      tags: [reports]
      summary: Create report run (Admin/PM session or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Report run created
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  report_run:
                    $ref: "#/components/schemas/ReportRun"
                required: [report_run]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/orgs/{org_id}/report-runs/{report_run_id}/publish:
    post:
      operationId: share_links__report_run_publish_post
      tags: [share_links]
      summary: Publish a report share link (Admin/PM session or API key write)
      security:
        - SessionCookie: []
        - ApiKeyBearer: []
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/ReportRunId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                expires_at:
                  type: string
                  format: date-time
                  nullable: true
      responses:
        "200":
          description: Share link published
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          description: Server error
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sessionid
      description: Django session cookie authentication.
    ApiKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: vrak_<prefix>.<secret>
      description: |
        API key authentication.
        Send `Authorization: Bearer vrak_<prefix>.<secret>`.
    GitLabWebhookToken:
      type: apiKey
      in: header
      name: X-Gitlab-Token
      description: GitLab webhook token configured per org integration.
  parameters:
    OrgId:
      name: org_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    Workflow:
      type: object
      additionalProperties: true
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        org_id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
      required: [id, org_id, name]
    WorkflowStage:
      type: object
      additionalProperties: true
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        workflow_id:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
        order:
          type: integer
        category:
          type: string
          enum: [backlog, in_progress, qa, done]
        progress_percent:
          type: integer
          minimum: 0
          maximum: 100
        is_done:
          type: boolean
        is_qa:
          type: boolean
        counts_as_wip:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, workflow_id, name, order, category, progress_percent, is_done]
    Project:
      type: object
      additionalProperties: true
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        org_id:
          $ref: "#/components/schemas/Uuid"
        workflow_id:
          allOf:
            - $ref: "#/components/schemas/Uuid"
          nullable: true
        progress_policy:
          type: string
          enum: [subtasks_rollup, workflow_stage]
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, org_id, name]
...